+++ src/systems/entity_control.hpp
@@ -13,5 +13,5 @@ void scroll_callback(GLFWwindow *window, double xoffset, double yoffset);
 
 void key_callback(GLFWwindow *window, int key, int scancode, int action, int mods);
 
-void entity_control(entt::registry &registry, entt::entity *cam, GLFWwindow *window, GLfloat deltaTime);
+void entity_control(entt::registry &registry, entt::entity *cam, GLFWwindow *window, double deltaTime);
